feat: 테이블뷰 셀의 데이터를 설정 #7
